fix for generation jobs
