+++ components/generator/cross-platform-generator-form.tsx
@@ -39,8 +39,6 @@ import {
   type Template,
 } from "@/components/templates/template-select";
 import { ProjectSelect } from "@/components/projects/project-select";
-import { RequirementRow, RequirementOption } from "@/types/requirements";
-
 import {
   type ModelKey,
   AI_MODELS,
@@ -54,29 +52,44 @@ import { toastWarning } from "@/lib/utils/toast-utils";
 
 // ─── Types ────────────────────────────────────────────────────────────────────
 
-type GenerationResult = { platform: string; count: number; error?: string };
-
-type CrossPlatformResponse = {
-  success?: boolean;
-  suite_id?: string;
-  total_test_cases?: number;
-  platforms?: string[];
-  generation_results?: GenerationResult[];
-  message?: string;
+// Response from POST /api/cross-platform-testing (job creation)
+type JobCreatedResponse = {
+  job_id?: string;
+  status?: string;
+  cases_requested?: number;
   error?: string;
   details?: string;
   upgradeRequired?: boolean;
   remaining?: number;
   requested?: number;
-  used?: number;
   limit?: number;
-  usage?: { requested: number; perPlatform: number; platforms: number };
 };
 
 type PlatformId = "web" | "mobile" | "api" | "accessibility" | "performance";
 type ApiProtocol = "REST" | "SOAP" | "GraphQL" | "gRPC" | "WebSocket";
 type ApiAuth = "None" | "Basic" | "Bearer" | "OAuth2" | "API Key" | "mTLS";
 type ApiFormat = "JSON" | "XML";
+
+type RequirementRow = {
+  id: string;
+  title: string;
+  description: string;
+  type: string;
+  priority: string;
+  status?: string;
+  project_id?: string | null;
+};
+
+type RequirementOption = {
+  id: string;
+  label: string;
+  title: string;
+  description: string;
+  type: string;
+  priority: string;
+  value: string;
+  project_id?: string | null;
+};
 
 type BootstrapResponse = { requirements: RequirementRow[] };
 
@@ -222,6 +235,13 @@ export function CrossPlatformGeneratorForm() {
   const { user, loading: authLoading } = useAuth();
 
   const [submitting, setSubmitting] = useState(false);
+  const [jobStatus, setJobStatus] = useState<{
+    jobId: string | null;
+    casesSaved: number;
+    casesRequested: number;
+    phase: "idle" | "queued" | "processing" | "done";
+  }>({ jobId: null, casesSaved: 0, casesRequested: 0, phase: "idle" });
+  const pollRef = React.useRef<ReturnType<typeof setInterval> | null>(null);
 
   // Core inputs
   const [title, setTitle] = useState("");
@@ -443,6 +463,79 @@ export function CrossPlatformGeneratorForm() {
     apiFormat,
   ]);
 
+  function stopPolling() {
+    if (pollRef.current) {
+      clearInterval(pollRef.current);
+      pollRef.current = null;
+    }
+  }
+
+  function startPolling(jobId: string, casesRequested: number) {
+    stopPolling();
+    setJobStatus({ jobId, casesSaved: 0, casesRequested, phase: "queued" });
+
+    pollRef.current = setInterval(async () => {
+      try {
+        const res = await fetch(`/api/jobs/${jobId}`, { cache: "no-store" });
+        if (!res.ok) return;
+        const data = (await res.json()) as {
+          status: string;
+          cases_saved?: number;
+          cases_requested?: number;
+          generation_id?: string;
+          partial?: boolean;
+          error?: string;
+        };
+
+        setJobStatus((prev) => ({
+          ...prev,
+          casesSaved: data.cases_saved ?? 0,
+          phase:
+            data.status === "pending"
+              ? "queued"
+              : data.status === "processing"
+                ? "processing"
+                : "done",
+        }));
+
+        if (data.status === "complete" || data.status === "failed") {
+          stopPolling();
+          setSubmitting(false);
+          setJobStatus({
+            jobId: null,
+            casesSaved: 0,
+            casesRequested: 0,
+            phase: "idle",
+          });
+
+          if (data.status === "failed") {
+            toast.error("Generation failed", {
+              description: data.error ?? "Please try again.",
+              duration: 8000,
+            });
+          } else {
+            toast.success("Cross-platform tests generated!", {
+              description: `Created ${data.cases_saved} test cases.`,
+              duration: 6000,
+            });
+            if (data.partial) {
+              toast.warning(
+                `${data.cases_saved} of ${casesRequested} cases generated — some batches failed. Try again for more.`,
+                { duration: 8000 },
+              );
+            }
+            router.push("/test-cases");
+          }
+        }
+      } catch {
+        // Network hiccup — keep polling
+      }
+    }, 3000);
+  }
+
+  // Clean up on unmount
+  React.useEffect(() => () => stopPolling(), []);
+
   const onSubmit = useCallback(
     async (e: React.FormEvent) => {
       e.preventDefault();
@@ -501,15 +594,12 @@ export function CrossPlatformGeneratorForm() {
           body: JSON.stringify(payload),
         });
 
-        const data = (await res.json()) as CrossPlatformResponse;
+        const data = (await res.json()) as JobCreatedResponse;
 
         if (res.status === 401) {
           toast.error("Please sign in to continue");
           router.push("/login");
-          return;
-        }
-        if ([502, 503, 504].includes(res.status)) {
-          toast.error("Generation timed out. Please try again.");
+          setSubmitting(false);
           return;
         }
 
@@ -534,22 +624,22 @@ export function CrossPlatformGeneratorForm() {
               },
             });
           }
+          setSubmitting(false);
           return;
         }
 
         if (!res.ok) {
           throw new Error(
-            data.details || data.error || `Failed (HTTP ${res.status})`,
+            data?.details || data?.error || `Failed (HTTP ${res.status})`,
           );
         }
 
-        toast.success("Cross-platform tests generated!", {
-          description:
-            data.message || `Created ${data.total_test_cases ?? 0} test cases.`,
-          duration: 6000,
-        });
+        if (!data.job_id) {
+          throw new Error("Server did not return a job ID. Please try again.");
+        }
 
-        router.push(data.suite_id ? "/test-library" : "/test-cases");
+        // Job created — start polling
+        startPolling(data.job_id, requestedTotal);
       } catch (err) {
         console.error("❌ Cross-platform generation error:", err);
         toast.error("Unable to generate cross-platform tests", {
@@ -557,7 +647,6 @@ export function CrossPlatformGeneratorForm() {
             err instanceof Error ? err.message : "Please try again later",
           duration: 8000,
         });
-      } finally {
         setSubmitting(false);
       }
     },
@@ -847,7 +936,9 @@ export function CrossPlatformGeneratorForm() {
                   </SelectContent>
                 </Select>
                 <p className="text-xs text-muted-foreground">
-                  Total Cases: {clampInt(perPlatformCount, 1, 20, 10)} ×{" "}
+                  AI generates a balanced mix of happy path, error handling,
+                  boundary, edge case, and security tests. Total ={" "}
+                  {clampInt(perPlatformCount, 1, 20, 10)} ×{" "}
                   {selectedPlatforms.length} platform(s) = {requestedTotal}{" "}
                   cases.
                 </p>
@@ -1054,20 +1145,30 @@ export function CrossPlatformGeneratorForm() {
             {/* Generation progress */}
             {submitting && (
               <div className="rounded-lg border bg-muted/30 p-4 space-y-3">
-                <div className="flex items-center gap-3">
-                  <Loader2 className="h-4 w-4 animate-spin text-primary shrink-0" />
-                  <span className="text-sm font-medium">
-                    Generating cross-platform test cases…
-                  </span>
+                <div className="flex items-center justify-between">
+                  <div className="flex items-center gap-3">
+                    <Loader2 className="h-4 w-4 animate-spin text-primary shrink-0" />
+                    <span className="text-sm font-medium">
+                      {jobStatus.phase === "queued"
+                        ? "Job queued — starting shortly…"
+                        : "Generating cross-platform test cases…"}
+                    </span>
+                  </div>
+                  {jobStatus.phase === "processing" &&
+                    jobStatus.casesRequested > 0 && (
+                      <span className="text-xs text-muted-foreground tabular-nums">
+                        {jobStatus.casesSaved} / {jobStatus.casesRequested}
+                      </span>
+                    )}
                 </div>
                 <div className="space-y-1.5">
                   <div className="h-1.5 w-full rounded-full bg-muted overflow-hidden">
                     <div className="h-1.5 rounded-full bg-primary animate-pulse w-full" />
                   </div>
                   <p className="text-xs text-muted-foreground">
-                    Running parallel AI batches across{" "}
-                    {selectedPlatforms.length} platform(s) — typically 20–60
-                    seconds.
+                    {jobStatus.phase === "queued"
+                      ? "Your generation job has been queued."
+                      : `Running parallel AI batches across ${selectedPlatforms.length} platform(s) — typically 20–60 seconds.`}
                   </p>
                 </div>
               </div>
